requirements.txt
# Core Dependencies
flask>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.2
pillow>=10.0.0  # For image processing

# Google Gemini AI
google-generativeai>=0.3.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
